Add archived appeal list

+++ card_maker_app/views/appeal.py
@@ -1,3 +1,4 @@
+from django.db.models import Q
 from rest_framework import viewsets, status
 from rest_framework.decorators import permission_classes, action
 from rest_framework.permissions import IsAdminUser
@@ -12,9 +13,23 @@ class AppealViewSet(viewsets.ModelViewSet):
     queryset = Appeal.objects.all()
     serializer_class = AppealSerializer
 
+    def list(self, request, *args, **kwargs):
+        appeals = Appeal.objects.filter(status='pending')
+
+        return Response(AppealSerializer(appeals, many=True).data)
+
+    @action(detail=False, methods=['get'])
+    def archived(self, request):
+        appeals = Appeal.objects.filter(~Q(status='pending'))
+
+        return Response(AppealSerializer(appeals, many=True).data)
+
     @action(detail=True, methods=['get'])
     def approve(self, request, pk):
         appeal = self.get_object()
+
+        if appeal.status != 'pending':
+            return Response({'detail': 'This appeal has already been processed'}, status.HTTP_400_BAD_REQUEST)
 
         appeal.status = 'approved'
         appeal.ban.delete()
@@ -27,6 +42,9 @@ class AppealViewSet(viewsets.ModelViewSet):
     @action(detail=True, methods=['get'])
     def decline(self, request, pk):
         appeal = self.get_object()
+
+        if appeal.status != 'pending':
+            return Response({'detail': 'This appeal has already been processed'}, status.HTTP_400_BAD_REQUEST)
 
         appeal.status = 'rejected'
         appeal.admin = request.user
